Hotfix product additional: missing record in cache

+++ odoo/local-src/product_additional/models/stock_pack_operation.py
@@ -19,9 +19,12 @@ class StockPackOperation(models.Model):
 
     @api.multi
     def unlink(self):
+        if not self:
+            return True
         if self.env.context.get('skip_additional'):
             return super(StockPackOperation, self).unlink()
 
+        moves_to_cancel = self.env['stock.move'].browse()
         for additional_move in self.mapped('additional_move_id'):
             ops = additional_move.mapped(
                 'linked_move_operation_ids.operation_id')
@@ -30,14 +33,21 @@ class StockPackOperation(models.Model):
             # Another one will come back if the main move is reassigned
             # but we don't want to loose what was done.
             if not ops_done:
-                super(StockPackOperation, ops).unlink()
-                ops_moves = ops.mapped('linked_move_operation_ids.move_id')
-                _logger.debug("Canceling additional moves %s",
-                              ops_moves.ids)
-                ops_moves.with_context(
-                    no_recompute_pack=True, force_cancel=True).action_cancel()
-                # A standard picker cannot delete a stock.move
-                ops_moves.sudo().unlink()
+                moves_to_cancel |= additional_move
+                ops_to_delete = ops - self
+                if ops_to_delete:
+                    super(StockPackOperation, ops_to_delete).unlink()
 
-        res = super(StockPackOperation, self).unlink()
-        return res
+        if moves_to_cancel:
+            _logger.debug("Canceling additional moves %s",
+                          moves_to_cancel.ids)
+            moves_to_cancel.with_context(
+                no_recompute_pack=True, force_cancel=True).action_cancel()
+            # A standard picker cannot delete a stock.move
+            #moves_to_cancel.sudo().with_context(
+            #    recompute=False,
+            #        ).unlink()
+
+        super(StockPackOperation, self).with_context(
+            skip_additional=True).unlink()
+        return True
